Add files via upload
Its a small project on a brand card from frontend mentor challenge. with the help of my HTML, CSS and JavaScript I have given some colour and beauty to the project. My CSS giving it the looks and making it look cool and the JavaScript  where I have coded my  function code by targeting the image id to enable the images to switch when it is clicked upon